third_party/WebKit/LayoutTests/http/tests/xmlhttprequest/response-text.html

<!DOCTYPE html>
<html>
<head>
</head>
<body>
    <script src="/js-test-resources/js-test.js"></script>
    <script>
    description("Request a file as the Text type with asynchronous XMLHttpRequest in a document context.");
    window.jsTestIsAsync = true;
    </script>
    <script src="resources/response-text.js"></script>
</body>
</html>
