third_party/blink/web_tests/platform/generic/http/tests/xmlhttprequest/response-text-expected.txt

Request a file as the Text type with asynchronous XMLHttpRequest in a document context.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS req.status is 200
PASS req.response is "PASS"
PASS req.status is 200
PASS req.response is "PASS"
PASS req.status is 200
PASS req.response is "PASS"
PASS successfullyParsed is true

TEST COMPLETE
